app.py
import os
import gradio as gr
from dotenv import load_dotenv
from deep_translator import GoogleTranslator
from languages import get_language_code, get_supported_languages, get_language_name

load_dotenv()

def translate_text_local(text, source_lang, target_lang):
    """Translate text using local deep-translator library (fast & accurate)."""
    
    if not text.strip():
        return ""
    
    try:
        # Handle auto-detection
        if source_lang == "Auto":
            # For auto-detection, we'll use Google's auto-detection
            target_code = get_language_code(target_lang)
            
            # Use Google Translator with auto-detection
            translator = GoogleTranslator(source='auto', target=target_code)
            result = translator.translate(text)
            
            # Get detected language info
            detected_lang = translator.detect_language(text)
            return f"[Auto-detected: {detected_lang}] {result}"
        else:
            # Get language codes
            source_code = get_language_code(source_lang)
            target_code = get_language_code(target_lang)
            
            # Validate languages
            if target_code == "auto":
                return "Error: Cannot translate to 'Auto' language"
            
            # Perform translation
            translator = GoogleTranslator(source=source_code, target=target_code)
            result = translator.translate(text)
            return result
            
    except Exception as e:
        print(f"❌ Translation error: {e}")
        return f"Translation error: {str(e)}"

def translate_text(text, source_lang, target_lang):
    """Main translation function."""
    return translate_text_local(text, source_lang, target_lang)

def swap_languages(source_lang, target_lang):
    """Swap source and target languages."""
    return target_lang, source_lang

def clear_texts():
    """Clear input and output text boxes."""
    return "", ""

# Create Gradio interface
def create_interface():
    """Create the Gradio interface."""
    with gr.Blocks(title="Multi-Language Translator (Local)", theme=gr.themes.Soft()) as interface:
        gr.Markdown("# 🌍 Multi-Language Translator (Local)")
        gr.Markdown("Translate text between 25+ languages using **local translation** - **zero latency, high accuracy!**")
        
        with gr.Row():
            source_lang = gr.Dropdown(
                choices=get_supported_languages(),
                value="Auto",
                label="Source Language",
                interactive=True
            )
            target_lang = gr.Dropdown(
                choices=[lang for lang in get_supported_languages() if lang != "Auto"],
                value="English",
                label="Target Language",
                interactive=True
            )
        
        input_text = gr.Textbox(
            lines=5,
            label="Input Text",
            placeholder="Enter text to translate...",
            interactive=True
        )
        
        output_text = gr.Textbox(
            lines=5,
            label="Translated Text",
            interactive=False
        )
        
        with gr.Row():
            translate_btn = gr.Button("Translate")
            swap_btn = gr.Button("🔄 Swap Languages")
            clear_btn = gr.Button("🗑️ Clear")
        
        gr.Markdown("---")
        gr.Markdown("**Note:** This application uses **local translation** powered by Google Translate - no internet required!")
        gr.Markdown("**Features:** ⚡ Zero latency • 🎯 High accuracy • 🔒 Privacy-focused • 🌐 Works offline")
        
        # Event handlers
        translate_btn.click(
            fn=translate_text,
            inputs=[input_text, source_lang, target_lang],
            outputs=output_text
        )
        
        swap_btn.click(
            fn=swap_languages,
            inputs=[source_lang, target_lang],
            outputs=[source_lang, target_lang]
        )
        
        clear_btn.click(
            fn=clear_texts,
            inputs=[],
            outputs=[input_text, output_text]
        )
        
        # Auto-translate on input change
        input_text.change(
            fn=translate_text,
            inputs=[input_text, source_lang, target_lang],
            outputs=output_text
        )
    
    return interface

if __name__ == "__main__":
    print("🚀 Starting Multi-Language Translator (Local Version)")
    print("✅ Using local translation - No API key required!")
    print("⚡ Zero latency • 🎯 High accuracy • 🔒 Privacy-focused")
    
    interface = create_interface()
    interface.launch(
        server_name="localhost",
        server_port=7860,
        share=False,
        show_error=True
    ) 