requirements.txt
gradio==4.44.0
deep-translator==1.11.4
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pydantic==2.5.0 